% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{changeterm}
\alias{changeterm}
\title{Change Term}
\usage{
changeterm(variable, overlap, term.type)
}
\arguments{
\item{variable}{A variable}

\item{overlap}{Another variable}

\item{term.type}{A type of term}
}
\description{
changeterm' defines the smallest term used for a class
}
\details{
This function returns a T/F logical statement and is used in the class assignment process
}
