% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{charter_ids}
\alias{charter_ids}
\title{Charter ID Data
A data frame with 121 rows and 11 columns}
\format{
## `charter_ids`
}
\source{
H:/Economists/Ed/KIDS/All Charter Schools/All/Charter IDs Master.csv
}
\usage{
charter_ids
}
\description{
Charter ID Data
A data frame with 121 rows and 11 columns
}
\keyword{datasets}
