% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{check1}
\alias{check1}
\title{Database Check 1}
\usage{
check1(expense)
}
\arguments{
\item{expense}{The file from the previous year}
}
\description{
'check1' allows us to run the first database check for charters and districts
}
\details{
This function is used in the charter and district database checks to confirm that the expense totals for each entity are correct.
}
