% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{check2}
\alias{check2}
\title{Database Check 2 Variables}
\usage{
check2(expense)
}
\arguments{
\item{expense}{The file from the previous year}
}
\description{
'check2' allows us to run the second database check for charters and districts
}
\details{
This function is used in the charter and district database checks to confirm that the treemap and school totals are correct. For our purposes, we want to make sure the expense totals match up when we look at them in different ways.
}
