% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{check3a}
\alias{check3a}
\title{Database Check 3a}
\usage{
check3a(expense, allocation)
}
\arguments{
\item{expense}{The expense file}

\item{allocation}{The allocation file}
}
\description{
'check3a' allows us to run the third database check for charters and districts
}
\details{
This function is used in the charter and district database checks to confirm that total expense is the same across expense and allocation files.
}
