% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{check3b}
\alias{check3b}
\title{Database Check 3a Variables}
\usage{
check3b(expense, allocation)
}
\arguments{
\item{expense}{The expense file}

\item{allocation}{The allocation file}
}
\description{
'check3b' allows us to run the first database check for charters and districts
}
\details{
This function is used in the charter and district database checks to confirm that teacher pay is the same across expense and allocation files.
}
