% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USBE_functions.R
\name{check_vars}
\alias{check_vars}
\title{Check Variables}
\usage{
check_vars(old_file, new_file)
}
\arguments{
\item{old_file}{The file from the previous year}

\item{new_file}{The file from the current year}
}
\description{
'check_vars' returns a table displaying a summary of the old and new datasets
}
\details{
This function is used in the USBE Quality Check stage of the update process to assess the new data against data that we have received in the past. Visually inspect the returned table for notable differences.
}
