% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{class_labels}
\alias{class_labels}
\title{CREATE CLASS LABELS
find the most frequent subject, course name, etc. per class id}
\usage{
class_labels(sm)
}
\arguments{
\item{sm}{The statewide merge dataframe for a given entity}
}
\description{
CREATE CLASS LABELS
find the most frequent subject, course name, etc. per class id
}
