% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{clength_recalc}
\alias{clength_recalc}
\title{CLENGTH RECALCUATION FOR DUPLICATE RECORDS FUNCTION
returns: START_STUDENT/END_STUDENT - new start/end dates
         CLENGTH_STUDENT - proportion of year spent in class
         NOTINCLASS_STUDENT - proportion of year between the start/end dates not spent in class}
\usage{
clength_recalc(starts, ends, clengths)
}
\description{
CLENGTH RECALCUATION FOR DUPLICATE RECORDS FUNCTION
returns: START_STUDENT/END_STUDENT - new start/end dates
         CLENGTH_STUDENT - proportion of year spent in class
         NOTINCLASS_STUDENT - proportion of year between the start/end dates not spent in class
}
