% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{confirm_charter_expense_time_function}
\alias{confirm_charter_expense_time_function}
\title{Confirm past files haven't been messed with}
\usage{
confirm_charter_expense_time_function()
}
\description{
'confirm_charter_expense_time_function' helps us see if files from past yearly updates have been accidentally altered in this round of yearly updates.
}
\details{
Because we are constantly changing and updating files, we want to make sure we aren't altering past updates and forgetting to update the correct files. This function focuses on charter expense.
}
