% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{district_allocation_columns}
\alias{district_allocation_columns}
\title{Confirm column numbers}
\usage{
district_allocation_columns(version)
}
\arguments{
\item{version}{This tells us which folder to pull from}
}
\description{
'district_allocation_columns' helps us see if files have the correct number of columns.
}
\details{
Because we are constantly changing and updating files, we want to make sure we have the correct number of columns so database uploads work properly.
}
