% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{district_expense}
\alias{district_expense}
\title{District Expense File Paths}
\usage{
district_expense(version)
}
\arguments{
\item{version}{This tells us which folder in the network drive we should be pulling from.}
}
\description{
'District Expense File Paths' allows us to make sure files are saved in the correct location in the network drive.
}
\details{
This function pulls the file path for each district and checks it against a threshold to make sure we're saving files in the correction location. If something's wrong, you may need to rerun the code or restore previous versions of old files.
}
