% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_checks.R
\name{district_final_database_checks}
\alias{district_final_database_checks}
\title{Database District Check}
\usage{
district_final_database_checks(version)
}
\arguments{
\item{version}{This tells us which folder we need to pull the data from}
}
\description{
'district_final_database_checks' allows us to run all of the database checks for all districts.
}
\details{
In an effort to expedite database checks, we run this script to systematically run the databse checks for each district and spit out the results This will mirror the results we find in the online database checks, but saves us the trouble of having to reupload the files to the cloud.
}
