% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_allocation.R
\name{.pullMatrix}
\alias{.pullMatrix}
\title{LEA ID}
\format{
## `lea_ids`
A data frame with 41 rows and 14 columns
\describe{
 \item{entity_name}{"Alpine School District"}
 \item{district_code}{"ALPS"}
 ...
 }
}
\usage{
.pullMatrix(lea)
}
\description{
LEA ID
}
