% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomatch.R
\name{get_nomatch}
\alias{get_nomatch}
\title{First: Create table from nomatch input df that is each check that needs to be done}
\usage{
get_nomatch(missing)
}
\description{
First: Create table from nomatch input df that is each check that needs to be done
}
