% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{getmode}
\alias{getmode}
\title{Get Mode}
\usage{
getmode(v, na.omit = T)
}
\arguments{
\item{v}{The column for which we are obtaining a mode value}

\item{na.omit}{Should NA values be excluded? Default is na.omit = T}
}
\description{
'getmode' returns the mode of a given column
}
\details{
This function returns the mode value for a given column
}
