% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{A parameter}

\item{y}{A parameter}
}
\description{
'notin' returns values that are not in the given parameters
}
\details{
This function is used throughout the PKids Update Process. It functions opposite to %in% and is equivalent to the function %nin% found in the hmisc package.
}
