% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{lea_ids}
\alias{lea_ids}
\title{LEA ID Data
A data frame with 41 rows and 14 columns
\describe{
 \item{entity_name}{"Alpine School District"}
 \item{district_code}{"ALPS"}
 \item{match}{"Alpine"}
 \item{add_type}{"County"}
 \item{district_name}{"Alpine School District"}
 \item{district_id}{"122"}
 \item{instritution_name}{"Alpine District"}
 \item{folder_name}{"Alpine School District (ALPS)"}
 \item{transparency_name}{"Alpine School District"}
 \item{transparency_id}{"259"}
 \item{lcode}{"1"}
 \item{district_location}{"99"}
 \item{transportation_id}{"01 Alpine"}
 \item{district_type}{"Urban"}
 }}
\format{
## `lea_ids`
}
\usage{
lea_ids
}
\description{
LEA ID Data
A data frame with 41 rows and 14 columns
\describe{
 \item{entity_name}{"Alpine School District"}
 \item{district_code}{"ALPS"}
 \item{match}{"Alpine"}
 \item{add_type}{"County"}
 \item{district_name}{"Alpine School District"}
 \item{district_id}{"122"}
 \item{instritution_name}{"Alpine District"}
 \item{folder_name}{"Alpine School District (ALPS)"}
 \item{transparency_name}{"Alpine School District"}
 \item{transparency_id}{"259"}
 \item{lcode}{"1"}
 \item{district_location}{"99"}
 \item{transportation_id}{"01 Alpine"}
 \item{district_type}{"Urban"}
 }
}
\keyword{datasets}
