% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter_allocation_scripts.R
\name{network_allocation}
\alias{network_allocation}
\title{Network Student Allocation}
\usage{
network_allocation()
}
\description{
'network_allocation' allows us to run all of the student allocation scripts in a loop for charter networks
}
\details{
Rather than running a dozen or so scripts manually, this code will process them in a loop and spit out a file with any errors.
}
