% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USBE_functions.R
\name{num_rows}
\alias{num_rows}
\title{Check Number of Rows}
\usage{
num_rows(old_file, new_file)
}
\arguments{
\item{old_file}{The file from the previous year}

\item{new_file}{The file from the current year}
}
\description{
Compares number of rows from past and current years
}
\details{
This function is used in the USBE Quality Check stage of the update process to assess the new data against data that we have received in the past. Inspect the number of rows by school year and note any substantial deviations from the old years to the new year.
}
