% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_allocation.R
\name{pullMatrix}
\alias{pullMatrix}
\title{Pull Student Allocation Matrix}
\usage{
pullMatrix(lea)
}
\arguments{
\item{lea}{The LEA code; input in quotes}
}
\description{
For use in Student Allocation
}
\details{
Load in specific entity's student allocation matrix
}
