% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charter_allocation_scripts.R
\name{single_allocation}
\alias{single_allocation}
\title{Database District Check}
\usage{
single_allocation()
}
\description{
'single_allocation' allows us to run all of the student allocation scripts in a loop for single charter schools
}
\details{
Rather than running a dozen or so scripts manually, this code will process them in a loop and spit out a file with any errors.
}
