% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{test_labels}
\alias{test_labels}
\title{Create Testing Labels}
\usage{
test_labels(sm)
}
\arguments{
\item{sm}{The statewide merge data frame for a given entity}
}
\description{
Create Testing Labels
}
\details{
Summarize testing data for visualization
}
