% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\docType{data}
\name{weekday_cal}
\alias{weekday_cal}
\title{# WEEKDAYS FUNCTION}
\format{
An object of class \code{Calendar} of length 17.
}
\usage{
weekday_cal
}
\description{
# WEEKDAYS FUNCTION
}
\keyword{datasets}
