% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USBE_functions.R
\name{which_vars}
\alias{which_vars}
\title{Which Variables}
\usage{
which_vars(old_file, new_file)
}
\arguments{
\item{old_file}{The file from the previous year}

\item{new_file}{The file from the current year}
}
\description{
Compares inclusion of variables from past and current years
}
\details{
This function is used in the USBE Quality Check stage of the update process to assess the new data against data that we have received in the past. Note variables included in the previous year that are not present in the current year, and vice versa.
}
