% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{changeterm}
\alias{changeterm}
\title{# FUNCTION TO DEFINE THE SMALLEST TERM USED FOR A CLASS
# RETURNS TRUE/FALSE LOGICAL STATEMENT}
\usage{
changeterm(variable, overlap, term.type)
}
\description{
# FUNCTION TO DEFINE THE SMALLEST TERM USED FOR A CLASS
# RETURNS TRUE/FALSE LOGICAL STATEMENT
}
