% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{class_labels}
\alias{class_labels}
\title{Create Class Labels}
\usage{
class_labels(sm)
}
\arguments{
\item{sm}{The statewide merge data frame for a given entity}
}
\description{
Create Class Labels
}
\details{
Find the most frequent subject, course name, etc. per class id
}
