% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{getmode}
\alias{getmode}
\title{# Get Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{The column in which we are obtaining a mode value}
}
\description{
'getmode' returns the mode of a given column
}
\details{
This function is used throughout the PKids Update Process.
}
