% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{lea_ids}
\alias{lea_ids}
\title{LEA ID Data
A data frame with 41 rows and 14 columns}
\format{
## `lea_ids`
}
\source{
H:/Economists/Ed/KIDS/All LEAs/District IDs Master.csv
}
\usage{
lea_ids
}
\description{
LEA ID Data
A data frame with 41 rows and 14 columns
}
\keyword{datasets}
