% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{new_startend}
\alias{new_startend}
\title{New Start/End}
\usage{
new_startend(start, end, n, nth)
}
\arguments{
\item{start}{The start}

\item{end}{The end}

\item{n}{The n}

\item{nth}{The nth}
}
\description{
new_startend' returns the new start and end dates for duplicated files
}
\details{
This function is used in defining classes in the class assignment process.
}
