% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{student_labels}
\alias{student_labels}
\title{CREATE STUDENT LABELS
find the most frequent school per student}
\usage{
student_labels(sm)
}
\arguments{
\item{sm}{The statewide merge dataframe for a given entity}
}
\description{
CREATE STUDENT LABELS
find the most frequent school per student
}
