% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\name{test_labels}
\alias{test_labels}
\title{CREATE TESTING LABELS ###
summarize testing data for visualization}
\usage{
test_labels(sm)
}
\description{
CREATE TESTING LABELS ###
summarize testing data for visualization
}
