% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_assignment_functions.R
\docType{data}
\name{weekday_cal}
\alias{weekday_cal}
\title{Weekdays}
\format{
An object of class \code{Calendar} of length 17.
}
\usage{
weekday_cal
}
\description{
'weekday_cal' returns a calendar object
}
\details{
This calendar is used primarily in the class assignment process when calculating the length of a school year and class length
}
\keyword{datasets}
